refactor: 디렉토리 계층 재구조화
