+++ src/components/InputTitle/InputTitle.stories.js
